// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "pages".
 */
export interface Page {
// ...
  expertises?: {
    enabled?: boolean | null;
    tag?: string | null;
    title?: string | null;
    content?: {
      root: {
        type: string;
        children: {
          type: any;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    } | null;
    button?: {
      href?: string | null;
      text?: string | null;
      icon?: ('arrow' | 'calendar' | 'estimate' | 'link') | null;
      border?: boolean | null;
    };
    expertisesList?:
      | {
          label?: string | null;
          icon?: (number | null) | Media;
          id?: string | null;
        }[]
      | null;
  };
// ...
}